improve: add avatar

+++ src/components/manager-card.tsx
@@ -15,6 +15,7 @@ import { toast } from "sonner"
 import { SessionProvider, useSession } from "next-auth/react"
 import { Link, Loader2, Unlink2 } from "lucide-react"
 import { api } from "@/trpc/react"
+import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar"
 
 type ManagerCardProps = {
   className?: string;
@@ -97,6 +98,10 @@ export function ManagerCard({
       return
     }
 
+    if (!session.user) {
+      return
+    }
+
     if (!id) {
       toast.error("Manager ID is required to unlink your FPL team.");
       return;
@@ -114,12 +119,27 @@ export function ManagerCard({
 
   return (
       <div className={cn("flex flex-col gap-6", className)}>
-        <Card className="text-white">
+        <Card className="text-white ">
           <CardHeader>
-            <CardTitle className="text-xl">{ name }</CardTitle>
-            <CardDescription>
-              { player_first_name } { player_last_name } { alias ? `/ @${alias}` : "" }
-            </CardDescription>
+            <div className="flex gap-2">
+              <div className="*:data-[slot=avatar]:ring-background flex -space-x-2 *:data-[slot=avatar]:ring-2 *:data-[slot=avatar]:grayscale">
+                <Avatar>
+                  <AvatarImage src={`https://placehold.co/20x20?text=${player_first_name[0]}${player_last_name[0]}`} alt="player_first_name" />
+                  <AvatarFallback>{player_first_name[0]}{player_last_name[0]}</AvatarFallback>
+                </Avatar>
+                <Avatar>
+                  <AvatarImage src={session?.user.image ?? `https://placehold.co/20x20?text=A`} alt={session?.user.name} />
+                  <AvatarFallback>{session?.user.name[0]}</AvatarFallback>
+                </Avatar>
+              </div>
+              <div>
+                <CardTitle className="text-xl">{ name }</CardTitle>
+                <CardDescription>
+                  { player_first_name } { player_last_name } { alias ? `/ @${alias}` : "" }
+                </CardDescription>
+
+              </div>
+            </div>
           </CardHeader>
             <CardContent>
               <div className="flex flex-col gap-4">
